backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai==0.3.2
